requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
python-dotenv==1.0.0
Werkzeug==2.3.7
email-validator==2.1.0
pandas==2.2.2
psycopg2-binary==2.9.9
SQLAlchemy==2.0.43
openpyxl==3.1.2
